Address review: split cursor stream lifecycle tests
Move proto half-close coverage into cursor_stream_test.go, dedupe
stream setup helpers, fail fast on unexpected post-close frames, and
extract shared unary-timeout stream context assertion.
